refactor(backend): add plain response to comply json, refactor ubc processing

+++ baggage-tracker-api/Services/UbcProcessor.cs
@@ -20,7 +20,7 @@ public class UbcProcessor(BaggageTrackerDbContext baggageTrackerDbContext)
 
         return new UniqueBaggageCode(
             flightNumber: fragments[0],
-            userId: int.Parse(fragments[1]),
+            userId: long.Parse(fragments[1]),
             baggageId: fragments[2].ParseAsGuid());
     }
